+++ APIBlox.AspNetCore/Filters/Action Filters/EnsureResponseResultActionFilter.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using APIBlox.NetCore.Extensions;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Logging;
@@ -29,10 +28,10 @@ namespace APIBlox.AspNetCore.Filters
             Func<object, object> ensureResponseCompliesWithAction
         )
         {
-            object DefaultFormat(object d) => new {Data = d};
+            object DefaultFormat(object d) => new { Data = d };
 
             _getsOnly = getsOnly;
-            _action = ensureResponseCompliesWithAction?? DefaultFormat;
+            _action = ensureResponseCompliesWithAction ?? DefaultFormat;
             _log = loggerFactory.CreateLogger<EnsureResponseResultActionFilter>();
         }
 
@@ -45,14 +44,14 @@ namespace APIBlox.AspNetCore.Filters
             var action = await next().ConfigureAwait(false);
             var result = action.Result as ObjectResult;
 
-            if (result?.Value is null || !(result.StatusCode is null)  && (result.StatusCode != StatusCodes.Status200OK && _getsOnly))
+            if (result?.Value is null || _getsOnly || result.StatusCode >= 300 || result.StatusCode < 200)
             {
                 var sc = result != null
                     ? result.StatusCode
                     : action.HttpContext.Response.StatusCode;
 
-                if (sc != 200)
-                    _log.LogInformation(() => $"Skipping execute as StatusCode is not 200, StatusCode is: {sc}");
+                if (sc >= 300 || sc < 200)
+                    _log.LogInformation(() => $"Skipping execute as StatusCode is: {sc}");
 
                 return;
             }
@@ -60,7 +59,7 @@ namespace APIBlox.AspNetCore.Filters
             var t = result.Value.GetType();
 
             ResultValueIsEnumerable = t.IsAssignableTo(typeof(IEnumerable)) && !t.IsAssignableTo(typeof(string));
-            ResultValueCount = ResultValueIsEnumerable ? ((IEnumerable<object>) result.Value).Count() : 0;
+            ResultValueCount = ResultValueIsEnumerable ? ((IEnumerable<object>)result.Value).Count() : 0;
 
             var retValue = _action(result.Value);